refactor(connections): extract selection UI components from connections.tsx
ConnectionGroupCard, BulkActionBar, and AddToAgentDialog were three
self-contained, prop-only components living inline in the 2045-line
connections.tsx. Move them into connection-selection-ui.tsx unchanged.
