+++ apps/mesh/src/web/routes/orgs/connections.tsx
@@ -5,7 +5,6 @@ import { CollectionTabs } from "@/web/components/collections/collection-tabs.tsx
 import { ConnectionCard } from "@/web/components/connections/connection-card.tsx";
 import { EmptyState } from "@/web/components/empty-state.tsx";
 import { ErrorBoundary } from "@/web/components/error-boundary";
-import { IntegrationIcon } from "@/web/components/integration-icon.tsx";
 import { Page } from "@/web/components/page";
 import type { RegistryItem } from "@/web/components/store/types";
 import { DeleteConnectionDialogs } from "@/web/components/delete-connection-dialogs";
@@ -82,7 +81,6 @@ import {
   useProjectContext,
   type ConnectionEntity,
   useVirtualMCPs,
-  type VirtualMCPEntity,
 } from "@decocms/mesh-sdk";
 import { useStudioTools } from "@/web/lib/studio-tools";
 import { toast } from "sonner";
@@ -135,10 +133,12 @@ import {
 // Grouping helpers (shared with agent add-connection dialog)
 // ---------------------------------------------------------------------------
 
+import { groupConnections } from "@/shared/utils/group-connections";
 import {
-  groupConnections,
-  type ConnectionGroup,
-} from "@/shared/utils/group-connections";
+  AddToAgentDialog,
+  BulkActionBar,
+  ConnectionGroupCard,
+} from "./connection-selection-ui.tsx";
 
 // ---------------------------------------------------------------------------
 // Connection type / status filter types
@@ -147,316 +147,6 @@ import {
 type ConnectionTypeFilter = "ALL" | "HTTP" | "SSE" | "Websocket" | "STDIO";
 
 type ConnectionStatusFilter = "ALL" | "active" | "inactive" | "error";
-
-// ---------------------------------------------------------------------------
-// Grouped card: collapsible row for connections sharing the same app_name
-// ---------------------------------------------------------------------------
-
-function ConnectionGroupCard({
-  group,
-  onOpen,
-  selectionMode,
-  selectedIds,
-  onToggleSelect,
-}: {
-  group: ConnectionGroup;
-  onOpen: () => void;
-  selectionMode: boolean;
-  selectedIds: Set<string>;
-  onToggleSelect: (id: string) => void;
-}) {
-  const allSelected = group.connections.every((c) => selectedIds.has(c.id));
-  const someSelected = group.connections.some((c) => selectedIds.has(c.id));
-
-  return (
-    <>
-      <ConnectionCard
-        connection={{
-          title: group.title,
-          icon: group.icon,
-          description: `${group.connections.length} instances`,
-        }}
-        onClick={() =>
-          selectionMode
-            ? (() => {
-                for (const c of group.connections) {
-                  if (allSelected) {
-                    if (selectedIds.has(c.id)) onToggleSelect(c.id);
-                  } else {
-                    if (!selectedIds.has(c.id)) onToggleSelect(c.id);
-                  }
-                }
-              })()
-            : onOpen()
-        }
-        className={cn(
-          selectionMode && allSelected && "ring-2 ring-primary",
-          selectionMode &&
-            someSelected &&
-            !allSelected &&
-            "ring-1 ring-primary/50",
-        )}
-        fallbackIcon={<Container />}
-        headerActionsAlwaysVisible
-        headerActions={
-          <div className="flex items-center gap-1">
-            {selectionMode ? (
-              <Checkbox
-                checked={
-                  allSelected ? true : someSelected ? "indeterminate" : false
-                }
-                onCheckedChange={() => {
-                  for (const c of group.connections) {
-                    if (allSelected) {
-                      if (selectedIds.has(c.id)) onToggleSelect(c.id);
-                    } else {
-                      if (!selectedIds.has(c.id)) onToggleSelect(c.id);
-                    }
-                  }
-                }}
-              />
-            ) : (
-              <div className="flex items-center gap-1">
-                <span className="text-xs text-muted-foreground font-normal">
-                  Connected
-                </span>
-                <span className="text-xs text-muted-foreground font-normal tabular-nums">
-                  x{group.connections.length}
-                </span>
-              </div>
-            )}
-            <div
-              className={cn(
-                "overflow-hidden transition-all duration-150 ease-out",
-                selectionMode
-                  ? "w-8 opacity-100"
-                  : "w-0 opacity-0 group-hover:w-8 group-hover:opacity-100",
-              )}
-            >
-              <DropdownMenu>
-                <DropdownMenuTrigger asChild>
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    className="h-8 w-8 p-0"
-                    onClick={(e) => e.stopPropagation()}
-                  >
-                    <DotsVertical size={20} />
-                  </Button>
-                </DropdownMenuTrigger>
-                <DropdownMenuContent
-                  align="end"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  <DropdownMenuItem
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      onOpen();
-                    }}
-                  >
-                    <Eye size={16} />
-                    Open
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </div>
-          </div>
-        }
-      />
-    </>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Floating bulk action bar (centered, same pattern as private-registry)
-// ---------------------------------------------------------------------------
-
-function BulkActionBar({
-  count,
-  total,
-  canManage,
-  canManageAgents,
-  onSelectAll,
-  onDeselectAll,
-  onDelete,
-  onAddToAgent,
-  onToggleStatus,
-  onCancel,
-}: {
-  count: number;
-  total: number;
-  canManage: boolean;
-  canManageAgents: boolean;
-  onSelectAll: () => void;
-  onDeselectAll: () => void;
-  onDelete: () => void;
-  onAddToAgent: () => void;
-  onToggleStatus: (status: "active" | "inactive") => void;
-  onCancel: () => void;
-}) {
-  if (count === 0) return null;
-
-  return (
-    <div className="fixed bottom-4 left-1/2 z-50 -translate-x-1/2">
-      <div className="rounded-xl border border-border bg-background/95 shadow-lg backdrop-blur px-3 py-2 flex items-center gap-2">
-        <div className="text-xs text-muted-foreground pr-1 tabular-nums">
-          {count} selected
-        </div>
-        {count < total ? (
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-7 text-xs px-2"
-            onClick={onSelectAll}
-          >
-            Select all ({total})
-          </Button>
-        ) : (
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-7 text-xs px-2"
-            onClick={onDeselectAll}
-          >
-            Clear selection
-          </Button>
-        )}
-        {canManageAgents && (
-          <Button
-            variant="outline"
-            size="sm"
-            className="h-7 text-xs px-2"
-            onClick={onAddToAgent}
-          >
-            <Plus size={13} />
-            Add to Agent
-          </Button>
-        )}
-        {canManage && (
-          <>
-            <Button
-              variant="outline"
-              size="sm"
-              className="h-7 text-xs px-2"
-              onClick={() => onToggleStatus("active")}
-            >
-              Enable
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              className="h-7 text-xs px-2"
-              onClick={() => onToggleStatus("inactive")}
-            >
-              Disable
-            </Button>
-            <Button
-              variant="destructive"
-              size="sm"
-              className="h-7 text-xs px-2"
-              onClick={onDelete}
-            >
-              <Trash01 size={13} />
-              Delete
-            </Button>
-          </>
-        )}
-        <Button
-          variant="ghost"
-          size="sm"
-          className="h-7 text-xs px-2"
-          onClick={onCancel}
-        >
-          <XClose size={13} />
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Add to Agent dialog
-// ---------------------------------------------------------------------------
-
-function AddToAgentDialog({
-  open,
-  onOpenChange,
-  agents,
-  onConfirm,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  agents: VirtualMCPEntity[];
-  onConfirm: (agentId: string) => void;
-}) {
-  const [selected, setSelected] = useState<string | null>(null);
-
-  return (
-    <Dialog open={open} onOpenChange={onOpenChange}>
-      <DialogContent className="sm:max-w-[400px]">
-        <DialogHeader>
-          <DialogTitle>Add to Agent</DialogTitle>
-          <DialogDescription>
-            Select an agent to add the selected connections to.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="max-h-60 overflow-auto py-2 space-y-1">
-          {agents.length === 0 ? (
-            <p className="text-sm text-muted-foreground text-center py-4">
-              No agents found
-            </p>
-          ) : (
-            agents.map((agent) => (
-              <button
-                key={agent.id}
-                type="button"
-                onClick={() => setSelected(agent.id)}
-                className={cn(
-                  "flex items-center gap-3 w-full rounded-md px-3 py-2 text-left transition-colors",
-                  selected === agent.id
-                    ? "bg-primary/10 ring-1 ring-primary"
-                    : "hover:bg-muted/50",
-                )}
-              >
-                <IntegrationIcon
-                  icon={agent.icon}
-                  name={agent.title}
-                  size="sm"
-                  className="shrink-0"
-                />
-                <div className="flex-1 min-w-0">
-                  <p className="text-sm font-medium truncate">{agent.title}</p>
-                  {agent.description && (
-                    <p className="text-xs text-muted-foreground truncate">
-                      {agent.description}
-                    </p>
-                  )}
-                </div>
-              </button>
-            ))
-          )}
-        </div>
-        <DialogFooter>
-          <Button variant="outline" onClick={() => onOpenChange(false)}>
-            Cancel
-          </Button>
-          <Button
-            disabled={!selected}
-            onClick={() => {
-              if (selected) {
-                onConfirm(selected);
-                onOpenChange(false);
-                setSelected(null);
-              }
-            }}
-          >
-            Add
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
 
 // ===========================================================================
